requirements.txt
asgiref==3.2.10
astroid==2.4.2
boto3==1.15.1
botocore==1.18.1
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1
django-allauth==0.42.0
django-countries==6.1.3
django-crispy-forms==1.9.2
django-phone-field==1.8.1
django-storages==1.10.1
gunicorn==20.0.4
idna==2.10
isort==4.3.21
jmespath==0.10.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
psycopg2-binary==2.8.6
pylint==2.5.3
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
stripe==2.51.0
toml==0.10.1
urllib3==1.25.10
wrapt==1.12.1
